+++ tests/eval_structure_generation/test_parser.py
@@ -5,13 +5,19 @@ from pathlib import Path
 from pytestarch.eval_structure_generation.file_import.config import Config
 from pytestarch.eval_structure_generation.file_import.file_filter import FileFilter
 from pytestarch.eval_structure_generation.file_import.parser import Parser
+from pytestarch.utils.partial_match_to_regex_converter import (
+    convert_partial_match_to_regex,
+)
 
 SOURCE_ROOT = Path(__file__).parent.parent
 RESOURCES_DIR = SOURCE_ROOT / "resources/importer"
 
 
 def test_parser_parses_all_files_in_directory() -> None:
-    parser = Parser(FileFilter(Config(("*__pycache__",))), SOURCE_ROOT)
+    parser = Parser(
+        FileFilter(Config([convert_partial_match_to_regex("*__pycache__")])),
+        SOURCE_ROOT,
+    )
 
     all_modules, parsed_modules = parser.parse(RESOURCES_DIR)