+++ docs/changelog.md
@@ -3,12 +3,14 @@
 This project uses semantic versioning and follows [keep a changelog](https://keepachangelog.com).
 
 
-## [Unreleased]
+## [1.4.0] -- 2023-04-09
 ### Added
+- Option to specify module names and file exclusions via regex.
 - Support for layer architecture.
 - Batch support for rule subjects.
 
 ### Changed
+- Complete module hierarchy present in graph, instead of only for modules that are part of an import dependency.
 - Error message only lists violated rule objects when multiple rule objects are defined.
 
 ## [1.3.1] -- 2023-01-27